Flytt typer som skal brukes til sanity mappe-struktur til egne filer.
