+++ src/schemas/baks/begrunnelse/ba-sak/typer.ts
@@ -4,75 +4,6 @@ import {
   EØSBegrunnelseDokumentNavn,
 } from '../../../../util/typer';
 
-export enum Begrunnelsestype {
-  INNVILGET = 'INNVILGET',
-  REDUKSJON = 'REDUKSJON',
-  AVSLAG = 'AVSLAG',
-  OPPHØR = 'OPPHØR',
-  FORTSATT_INNVILGET = 'FORTSATT_INNVILGET',
-  ENDRET_UTBETALINGSPERIODE = 'ENDRET_UTBETALINGSPERIODE',
-  ETTER_ENDRET_UTBETALINGSPERIODE = 'ETTER_ENDRET_UTBETALINGSPERIODE',
-}
-
-export const begrunnelsestyperTilMenynavn: Record<
-  Begrunnelsestype,
-  { title: string; value: Begrunnelsestype }
-> = {
-  INNVILGET: { title: 'Innvilget', value: Begrunnelsestype.INNVILGET },
-  REDUKSJON: { title: 'Reduksjon', value: Begrunnelsestype.REDUKSJON },
-  AVSLAG: { title: 'Avslag', value: Begrunnelsestype.AVSLAG },
-  OPPHØR: { title: 'Opphør', value: Begrunnelsestype.OPPHØR },
-  FORTSATT_INNVILGET: { title: 'Fortsatt innvilget', value: Begrunnelsestype.FORTSATT_INNVILGET },
-  ENDRET_UTBETALINGSPERIODE: {
-    title: 'Endret utbetaling',
-    value: Begrunnelsestype.ENDRET_UTBETALINGSPERIODE,
-  },
-  ETTER_ENDRET_UTBETALINGSPERIODE: {
-    title: 'Etter endret utbetaling',
-    value: Begrunnelsestype.ETTER_ENDRET_UTBETALINGSPERIODE,
-  },
-};
-
-export enum Valgbarhet {
-  STANDARD = 'STANDARD',
-  AUTOMATISK = 'AUTOMATISK',
-  TILLEGGSTEKST = 'TILLEGGSTEKST',
-}
-
-export const valgbarhetTilMenynavn: Record<
-  Valgbarhet,
-  {
-    title: string;
-    value: Valgbarhet;
-  }
-> = {
-  AUTOMATISK: { title: 'Automatisk', value: Valgbarhet.AUTOMATISK },
-  STANDARD: { title: 'Standard', value: Valgbarhet.STANDARD },
-  TILLEGGSTEKST: { title: 'Tilleggstekst', value: Valgbarhet.TILLEGGSTEKST },
-};
-
-export enum BegrunnelseTema {
-  NASJONAL = 'NASJONAL',
-  FELLES = 'FELLES',
-  PRIMÆRLAND = 'PRIMÆRLAND',
-  SEKUNDÆRLAND = 'SEKUNDÆRLAND',
-  INSTITUSJON = 'INSTITUSJON',
-}
-
-export const begrunnelseTemaTilMenynavn: Record<
-  BegrunnelseTema,
-  {
-    title: string;
-    value: BegrunnelseTema;
-  }
-> = {
-  FELLES: { title: 'Felles', value: BegrunnelseTema.FELLES },
-  INSTITUSJON: { title: 'Institusjon', value: BegrunnelseTema.INSTITUSJON },
-  NASJONAL: { title: 'Nasjonal', value: BegrunnelseTema.NASJONAL },
-  PRIMÆRLAND: { title: 'Primærland', value: BegrunnelseTema.PRIMÆRLAND },
-  SEKUNDÆRLAND: { title: 'Sekundærland', value: BegrunnelseTema.SEKUNDÆRLAND },
-};
-
 export const hjemler = ['2', '4', '5', '9', '10', '11', '12', '14', '17', '18', '22'];
 
 export const hjemlerFolketrygdloven = ['2-5', '2-8'];
