+++ src/util/typer.ts
@@ -53,9 +53,10 @@ export enum SanityTyper {
 
 export enum BegrunnelseDokumentNavn {
   BA_BEGRUNNELSE = 'begrunnelse',
-  BEGRUNNELSE_TYPE = 'begrunnelsetype',
+  RESULTAT = 'resultat',
   VALGBARHET = 'valgbarhet',
   TEMA = 'tema',
+  BEGRUNNELSE_TYPE = 'begrunnelsetype',
   HJEMLER = 'hjemler',
   HJEMLER_EØS_FORORDNINGEN_833 = 'hjemlerEOSForordningen883',
   HJEMLER_EØS_FORORDNINGEN_987 = 'hjemlerEOSForordningen987',
